create image socrol view and change some style and prepare it as component #12
